refactor(ai-builder): Consolidate pairwise evaluation code into dedicated folder
Move all pairwise-related files into evaluations/pairwise/ directory
since pairwise evaluation supports both local and LangSmith modes,
making the previous langsmith/ location confusing.

Files moved:
- langsmith/pairwise-runner.ts → pairwise/runner.ts
- langsmith/pairwise-generator.ts → pairwise/generator.ts
- langsmith/pairwise-metrics-builder.ts → pairwise/metrics-builder.ts
- chains/pairwise-judge-chain.ts → pairwise/judge-chain.ts
- utils/judge-panel.ts → pairwise/judge-panel.ts
- types/pairwise.ts → pairwise/types.ts
