+++ packages/@n8n/ai-workflow-builder.ee/evaluations/README.md
@@ -16,10 +16,16 @@ evaluations/
 ├── cli/                 # CLI evaluation implementation
 │   ├── runner.ts       # Main CLI evaluation orchestrator
 │   └── display.ts      # Console output and progress tracking
-├── langsmith/          # Langsmith integration
+├── langsmith/          # Langsmith integration (non-pairwise)
 │   ├── evaluator.ts    # Langsmith-compatible evaluator function
-│   ├── runner.ts       # Langsmith evaluation orchestrator
-│   └── pairwise-runner.ts # Pairwise evaluation orchestrator
+│   └── runner.ts       # Langsmith evaluation orchestrator
+├── pairwise/           # Pairwise evaluation (local + LangSmith modes)
+│   ├── runner.ts       # Orchestration for both local and LangSmith modes
+│   ├── generator.ts    # Target function + workflow generation
+│   ├── metrics-builder.ts # LangSmith metric builders
+│   ├── judge-chain.ts  # LLM judge chain
+│   ├── judge-panel.ts  # Multi-judge execution and aggregation
+│   └── types.ts        # Dataset input/output types
 ├── core/               # Shared evaluation logic
 │   ├── environment.ts  # Test environment setup and configuration
 │   └── test-runner.ts  # Core test execution logic
@@ -30,7 +36,6 @@ evaluations/
 ├── chains/             # LLM evaluation chains
 │   ├── test-case-generator.ts    # Dynamic test case generation
 │   ├── workflow-evaluator.ts     # Main orchestrator for parallel evaluation
-│   ├── pairwise-judge-chain.ts   # Pairwise do/don't criteria evaluator (LLM judge)
 │   └── evaluators/               # Individual metric evaluators
 │       ├── index.ts              # Evaluator exports
 │       ├── functionality-evaluator.ts      # Functional correctness evaluation
@@ -104,15 +109,15 @@ The Langsmith integration provides two key components:
 
 #### 6. Pairwise Evaluation
 
-Pairwise evaluation provides a criteria-based approach to workflow evaluation with hierarchical scoring and multi-judge consensus. It evaluates workflows against a custom set of "do" and "don't" rules defined in the dataset.
+Pairwise evaluation provides a criteria-based approach to workflow evaluation with hierarchical scoring and multi-judge consensus. It evaluates workflows against a custom set of "do" and "don't" rules defined in the dataset. All pairwise-related code is consolidated in the `pairwise/` directory.
 
-**Judge Chain (`chains/pairwise-judge-chain.ts`):**
+**Judge Chain (`pairwise/judge-chain.ts`):**
 - Evaluates workflows against a checklist of criteria (dos and don'ts)
 - Uses an LLM to determine if each criterion passes or fails
 - Requires evidence-based justification for each decision
 - Returns `primaryPass` (true only if ALL criteria pass) and `diagnosticScore` (ratio of passes)
 
-**Runner (`langsmith/pairwise-runner.ts`):**
+**Runner (`pairwise/runner.ts`):**
 - Generates workflows from prompts in the dataset
 - Runs multiple LLM judges in parallel for each evaluation (configurable via `--judges`)
 - Aggregates judge results using majority vote
