+++ packages/@n8n/ai-workflow-builder.ee/evaluations/pairwise/judge-chain.test.ts
@@ -1,8 +1,8 @@
 import type { BaseChatModel } from '@langchain/core/language_models/chat_models';
 
+import { evaluateWorkflowPairwise, type PairwiseEvaluationInput } from './judge-chain';
 import type { SimpleWorkflow } from '../../src/types/workflow';
 import * as baseEvaluator from '../chains/evaluators/base';
-import { evaluateWorkflowPairwise, type PairwiseEvaluationInput } from './pairwise-judge-chain';
 
 // Mock the base evaluator module
 jest.mock('../chains/evaluators/base', () => ({
