+++ packages/@n8n/ai-workflow-builder.ee/evaluations/pairwise/judge-panel.ts
@@ -1,10 +1,7 @@
 import type { BaseChatModel } from '@langchain/core/language_models/chat_models';
 
+import { evaluateWorkflowPairwise, type PairwiseEvaluationResult } from './judge-chain';
 import type { SimpleWorkflow } from '../../src/types/workflow';
-import {
-	evaluateWorkflowPairwise,
-	type PairwiseEvaluationResult,
-} from '../chains/pairwise-judge-chain';
 
 // ============================================================================
 // Types
